test/integration/framework: cleanup master_utils.go
